Time: 110 ms (49.46%), Space: 41.9 MB (57.42%) - LeetHub

+++ 881-boats-to-save-people/881-boats-to-save-people.cpp
@@ -6,9 +6,9 @@ public:
         long long int ans = 0;
         while (a<=b) {
             ans++;
-            if (people[a] + people[b] <= limit)
+            if (people[a]+people[b]<=limit)
                 a++;
             b--;
-        }return ans;
+        }return (int)ans;
     }
 };
